TST/REF: collect tests by method from tests.internals (#37420)
